[CPDLP-1564] - add specs for queries
