+++ app/services/finance/npq/assurance_report/query.rb
@@ -19,6 +19,7 @@ module Finance
         def sql
           <<~EOSQL
             SELECT
+              pd.id                   AS id,
               pi.external_identifier  AS participant_id,
               u.full_name             AS participant_name,
               tp.trn                  AS trn,
